Refactor some shared logic out of SynapseE2E and SynapseExtension E2E
