+++ core/src/test/scala/com/microsoft/azure/synapse/ml/nbtest/SynapseTests.scala
@@ -3,16 +3,13 @@
 
 package com.microsoft.azure.synapse.ml.nbtest
 
-import com.microsoft.azure.synapse.ml.build.BuildInfo
 import com.microsoft.azure.synapse.ml.core.env.FileUtilities
 import com.microsoft.azure.synapse.ml.core.test.base.TestBase
 import com.microsoft.azure.synapse.ml.io.http.RESTHelpers.sendAndParseJson
 import com.microsoft.azure.synapse.ml.nbtest.SynapseUtilities._
-import org.apache.commons.io.FileUtils
 import org.apache.http.client.methods.HttpGet
 
 import java.io.File
-import java.lang.ProcessBuilder.Redirect
 import java.util.concurrent.TimeUnit
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
@@ -43,66 +40,9 @@ class SynapseTestCleanup extends TestBase {
 }
 
 class SynapseTests extends TestBase {
+  SharedNotebookE2ETestUtilities.generateNotebooks()
 
-  val resourcesDirectory = new File(getClass.getResource("/").toURI)
-  val notebooksDir = new File(resourcesDirectory, "generated-notebooks")
-  FileUtils.deleteDirectory(notebooksDir)
-  assert(notebooksDir.mkdirs())
-
-  val notebooks: Array[File] = FileUtilities.recursiveListFiles(FileUtilities
-    .join(BuildInfo.baseDirectory.getParent, "notebooks/features")
-    .getCanonicalFile)
-    .filter(_.getName.endsWith(".ipynb"))
-    .map { f =>
-      FileUtilities.copyFile(f, notebooksDir, true)
-      val newFile = new File(notebooksDir, f.getName)
-      val targetName = new File(notebooksDir, f.getName.replace(" ", "").replace("-", ""))
-      newFile.renameTo(targetName)
-      targetName
-    }
-
-  assert(notebooks.length > 1)
-
-  def isWindows: Boolean = {
-    sys.props("os.name").toLowerCase.contains("windows")
-  }
-
-  def osPrefix: Seq[String] = {
-    if (isWindows) {
-      Seq("cmd", "/C")
-    } else {
-      Seq()
-    }
-  }
-
-  def runCmd(cmd: Seq[String],
-             wd: File = new File("."),
-             envVars: Map[String, String] = Map()): Unit = {
-    val pb = new ProcessBuilder()
-      .directory(wd)
-      .command(cmd: _*)
-      .redirectError(Redirect.INHERIT)
-      .redirectOutput(Redirect.INHERIT)
-    val env = pb.environment()
-    envVars.foreach(p => env.put(p._1, p._2))
-    assert(pb.start().waitFor() == 0)
-  }
-
-  def condaEnvName: String = "synapseml"
-
-  def activateCondaEnv: Seq[String] = {
-    if (sys.props("os.name").toLowerCase.contains("windows")) {
-      osPrefix ++ Seq("activate", condaEnvName, "&&")
-    } else {
-      Seq()
-      //TODO figure out why this doesent work
-      //Seq("/bin/bash", "-l", "-c", "source activate " + condaEnvName, "&&")
-    }
-  }
-
-  runCmd(activateCondaEnv ++ Seq("jupyter", "nbconvert", "--to", "python", "*.ipynb"), notebooksDir)
-
-  val selectedPythonFiles: Array[File] = FileUtilities.recursiveListFiles(notebooksDir)
+  val selectedPythonFiles: Array[File] = FileUtilities.recursiveListFiles(SharedNotebookE2ETestUtilities.NotebooksDir)
     .filter(_.getAbsolutePath.endsWith(".py"))
     .filterNot(_.getAbsolutePath.contains("HyperParameterTuning"))
     .filterNot(_.getAbsolutePath.contains("VowpalWabbitOverview"))
