+++ core/src/test/scala/com/microsoft/azure/synapse/ml/nbtest/SynapseUtilities.scala
@@ -5,10 +5,10 @@ package com.microsoft.azure.synapse.ml.nbtest
 
 import com.microsoft.azure.synapse.ml.Secrets
 import com.microsoft.azure.synapse.ml.build.BuildInfo
-import com.microsoft.azure.synapse.ml.core.env.FileUtilities
 import com.microsoft.azure.synapse.ml.io.http.RESTHelpers
 import com.microsoft.azure.synapse.ml.io.http.RESTHelpers.{safeSend, sendAndParseJson}
 import com.microsoft.azure.synapse.ml.nbtest.SynapseUtilities._
+import com.microsoft.azure.synapse.ml.nbtest.SharedNotebookE2ETestUtilities._
 import org.apache.http.client.entity.UrlEncodedFormEntity
 import org.apache.http.client.methods._
 import org.apache.http.entity.StringEntity
@@ -21,7 +21,6 @@ import scala.annotation.tailrec
 import scala.collection.JavaConverters._
 import scala.concurrent.{ExecutionContext, Future, blocking}
 import scala.language.postfixOps
-import scala.sys.process._
 
 case class LivyBatchData(id: Int,
                          state: String,
@@ -143,17 +142,6 @@ object SynapseUtilities {
     s"$WorkspaceRoot/livyApi/versions/2019-11-01-preview/sparkPools/$sparkPool/batches"
   }
 
-  def listPythonJobFiles(): Seq[String] = {
-    val rootDirectory = FileUtilities
-      .join(BuildInfo.baseDirectory.getParent, "notebooks/features")
-      .getCanonicalFile
-
-    FileUtilities.recursiveListFiles(rootDirectory)
-      .filter(_.getAbsolutePath.endsWith(".py"))
-      .map(file => file.getAbsolutePath)
-      .sorted
-  }
-
   def getQueuedJobs(poolName: String): Applications = {
     val uri: String =
       WorkspaceRoot +
@@ -224,14 +212,6 @@ object SynapseUtilities {
     createRequest.setHeader("Authorization", s"Bearer $SynapseToken")
     createRequest.setEntity(new StringEntity(livyPayload))
     LivyBatch(sendAndParseJson(createRequest).convertTo[LivyBatchData], runName, poolName)
-  }
-
-  def exec(command: String): String = {
-    val os = sys.props("os.name").toLowerCase
-    os match {
-      case x if x contains "windows" => Seq("cmd", "/C") ++ Seq(command) !!
-      case _ => command !!
-    }
   }
 
   private def bigDataPoolBicepPayload(bigDataPoolName: String,
